Keywords/requirements.txt
dnspython
google-cloud-language
google-cloud-storage
google-cloud-speech
pymongo